refactor(ci-scripts): dedupe the buffered leg runner into spawn-tagged.mjs
perf-coverage-fanout.mjs, property-fanout.mjs and chdb-roundtrip.mjs each
hand-rolled an identical buffer-then-print-whole runLeg. Add runLegBuffered
to lib/spawn-tagged.mjs alongside the existing streamed runners and have
all three fan-out scripts import it instead of keeping their own copy.
