+++ .github/scripts/lib/spawn-tagged.mjs
@@ -1,13 +1,17 @@
 // spawn-tagged.mjs — run child processes concurrently with readable output.
 //
-// Shared by the two regeneration runners that fan a corpus walk out across
-// processes: golden-update.mjs (the TXTAR goldens, partitioned by
-// test/spec/shard.go) and cardinality-baseline-update.mjs (the perf cardinality
-// baseline, partitioned by test/perf/profile/shard.go). Both need exactly the
-// same two things — tag each child's output so concurrent legs stay readable,
-// and let every leg finish before deciding whether the group failed — and a
-// second copy of that would be a second set of failure semantics to keep in
-// step.
+// Shared by every fan-out runner that splits a corpus walk across concurrent
+// `go test` processes: golden-update.mjs and cardinality-baseline-update.mjs
+// tag each child's output and stream it line-by-line (`runTagged` /
+// `runAllTagged`); perf-coverage-fanout.mjs, property-fanout.mjs and
+// chdb-roundtrip.mjs instead buffer each child's output whole and print it
+// after every leg finishes (`runLegBuffered`), because their failures are
+// multi-line blocks — a fixture diff or a goroutine dump — that streaming
+// would interleave into unreadable garbage across concurrent legs. All five
+// callers need the same failure-propagation contract regardless of which
+// output mode they use — let every leg finish before deciding whether the
+// group failed — and a second copy of that per caller would be a second set
+// of failure semantics to keep in step.
 //
 // No env inputs; it is a library, not a step.
 
@@ -65,4 +69,36 @@ export async function runAllTagged(commands, cwd, sink) {
   );
 
   return codes.find((c) => c !== 0) ?? 0;
+}
+
+/**
+ * Runs one leg (`{ argv, env? }`) to completion, buffering its stdout and
+ * stderr rather than streaming them.
+ *
+ * Buffered rather than tagged-and-streamed because these callers' legs run
+ * concurrently and a failure is a multi-line block — a `go test` fixture
+ * diff or a goroutine dump — that becomes unreadable once several legs
+ * shuffle their lines into each other. Callers print each leg's whole
+ * buffer only after every leg has finished. Resolves rather than rejects
+ * (including on a spawn error) so `Promise.all` over several legs always
+ * waits for the group, the same contract `runTagged` holds for the
+ * streaming callers above.
+ */
+export function runLegBuffered(leg) {
+  return new Promise((resolve) => {
+    const [cmd, ...args] = leg.argv;
+    const child = spawn(cmd, args, {
+      env: { ...process.env, ...leg.env },
+      stdio: ['ignore', 'pipe', 'pipe'],
+    });
+    let out = '';
+    child.stdout.on('data', (d) => {
+      out += d;
+    });
+    child.stderr.on('data', (d) => {
+      out += d;
+    });
+    child.on('error', (err) => resolve({ leg, code: 1, out: `${out}\nspawn ${cmd}: ${err.message}` }));
+    child.on('close', (code) => resolve({ leg, code: code ?? 1, out }));
+  });
 }
